feat: migrate appView.js
