+++ migration-to-ts/src/components/view/appView.ts
@@ -1,0 +1,26 @@
+import { INews } from '../../types/types';
+import News from './news/news';
+import Sources from './sources/sources';
+
+export class AppView {
+  news: News;
+
+  sources: Sources;
+
+  constructor() {
+    this.news = new News();
+    this.sources = new Sources();
+  }
+
+  drawNews(data: INews): void {
+    const values = data?.articles ? data?.articles : [];
+    this.news.draw(values);
+  }
+
+  drawSources(data: INews): void {
+    const values = data?.sources ? data?.sources : [];
+    this.sources.draw(values);
+  }
+}
+
+export default AppView;
